fix - reservation variables and day counter funcion
