+++ package.json
@@ -22,6 +22,7 @@
     "dotenv": "^14.3.2",
     "express": "^4.17.2",
     "jsonwebtoken": "^8.5.1",
+    "moment": "^2.29.1",
     "morgan": "^1.10.0",
     "nodemailer": "^6.7.2",
     "pg": "^8.7.1",
