--- prueba.js
@@ -1,4 +1,0 @@
-const f1 = new Date("2021-02-01")
-const f2 = new Date("2021-02-02")
-
-f1>f2? console.log('si'):console.log('no');
